ml-models/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
typing-extensions